* common-src/amanda.h: use ssize_t in writev return value 	* common-src/bsd-security.c (bsd_stream_write,bsd_stream_read): 	use ssize_t over int 	* common-src/file.c (areads): use ssize_t to hold read() return val 	* common-src/fileheader.c (fill_buffer): usse ssize_t to hold read() 	return val 	* common-src/security.h: stream_read's callback now takes a ssize_t 	as the third argument. 	* common-src/writev.c (writev): return a ssize_t

+++ common-src/amanda.h
@@ -24,7 +24,7 @@
  * file named AUTHORS, in the root directory of this distribution.
  */
 /*
- * $Id: amanda.h,v 1.73 1998/12/14 20:23:00 kashmir Exp $
+ * $Id: amanda.h,v 1.74 1999/03/04 22:39:58 kashmir Exp $
  *
  * the central header file included by all amanda sources
  */
@@ -1105,7 +1105,7 @@ extern pid_t waitpid P((pid_t pid, amwait_t *stat_loc, int options));
 #endif
 
 #ifndef HAVE_WRITEV_DECL
-extern int writev P((int fd, const struct iovec *iov, int iovcnt));
+extern ssize_t writev P((int fd, const struct iovec *iov, int iovcnt));
 #endif
 
 #ifndef STDIN_FILENO
